pyproject.toml
[project]
name = "vault-migrate"
version = "0.1.0"
description = "Migrate an Obsidian vault to a SQLite graph database for AI agent workflows"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0.3",
]

[project.scripts]
vault-db = "vault_db.cli:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["vault_db"]
